third_party/blink/web_tests/dom/domparsing/xmlserializer-doctype.html

<html>
<head>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script>
        test(() => {
            var docType = window.document.implementation.createDocumentType("aDocTypeName", "aPublicID", "aSystemID");
            var doc = window.document.implementation.createDocument("", "", docType);

            var serializer = new XMLSerializer();

            var result = serializer.serializeToString(docType);
            assert_equals(result, '<!DOCTYPE aDocTypeName PUBLIC "aPublicID" "aSystemID">', 'The DocumentType node should serialize to "' + result + '"');
        }, "This tests XMLSerializer.serializeToString() on a DocumentType node that has a document associated with it");
    </script>
</head>
<body>
</body>
</html>
